requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0

# Vector search and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# LLM providers
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
tiktoken==0.5.2

# API and web
fastapi==0.108.0
uvicorn[standard]==0.25.0
streamlit==1.29.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Deduplication and hashing
datasketch==1.6.4

# Logging
structlog==23.3.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
